requirements.txt
asgiref==3.2.3
astroid==2.3.3
Django==3.0.2
django-ckeditor==5.8.0
django-crispy-forms==1.8.1
django-filter==2.2.0
django-filters==0.2.1
django-js-asset==1.2.2
django-search-listview==0.2.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.761
mypy-extensions==0.4.3
Pillow==7.0.0
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
wrapt==1.11.2
